+++ src/cpusim/mt_engine.h
@@ -11,12 +11,12 @@ namespace CPUSIM
 
         MT_ENGINE(CORE::BODY_STATE_VEC body_states_ic,
                   CORE::DT dt,
-                  int n_thread,
+                  size_t n_thread,
                   std::optional<std::string> body_states_log_dir_opt = {});
 
         virtual CORE::BODY_STATE_VEC execute(int n_iter) override;
 
     private:
-        int n_thread_;
+        size_t n_thread_;
     };
 }
